+++ SDeMo/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.4.0`
+
+- **added** support `Bagging` models as a component in an `Ensemble`
+- **added** support for null classifiers based on `Bagging` models
+- **added** variable selection for `Bagging` models [#405]
+- **added** a more general syntax for variable selection (`VariableSelectionStrategy`)
+- **added** a `ChainedTransform` type to chain two data transformation steps, useful for PCA + z-score [#408]
+- **fixed** the issue with variable selection reseting the model variables [#400]
+- **added** a method for `noselection!` where no folds are given
+
 ## `v1.3.4`
 
 - **improved** the memory allocation of Logistic regression